+++ Tests/KeyboardKitTests/Actions/KeyboardActionTests.swift
@@ -79,10 +79,6 @@ class KeyboardActionTests: QuickSpec {
         
         describe("is input action") {
             
-            func result(for action: KeyboardAction) -> Bool {
-                action.isInputAction
-            }
-            
             it("is true for some actions") {
                 expected = [
                     .character(""),
@@ -90,32 +86,33 @@ class KeyboardActionTests: QuickSpec {
                     .image(description: "", keyboardImageName: "", imageName: ""),
                     .space,
                     .systemImage(description: "", keyboardImageName: "", imageName: "")]
-                expected.forEach { expect(result(for: $0)).to(beTrue()) }
-                unexpected.forEach { expect(result(for: $0)).to(beFalse()) }
+                expected.forEach { expect($0.isInputAction).to(beTrue()) }
+                unexpected.forEach { expect($0.isInputAction).to(beFalse()) }
             }
         }
         
-        describe("is input action") {
+        describe("is primary action") {
             
-            func result(for action: KeyboardAction) -> Bool {
-                action.isShift
+            it("is true for some actions") {
+                expected = [.done, .go, .ok, .search]
+                expected.forEach { expect($0.isPrimaryAction).to(beTrue()) }
+                unexpected.forEach { expect($0.isPrimaryAction).to(beFalse()) }
             }
+        }
+        
+        describe("is shift") {
             
             it("is only true for all shift actions") {
                 expected = [
                     .shift(currentState: .capsLocked),
                     .shift(currentState: .lowercased),
                     .shift(currentState: .uppercased)]
-                expected.forEach { expect(result(for: $0)).to(beTrue()) }
-                unexpected.forEach { expect(result(for: $0)).to(beFalse()) }
+                expected.forEach { expect($0.isShift).to(beTrue()) }
+                unexpected.forEach { expect($0.isShift).to(beFalse()) }
             }
         }
         
         describe("is system action") {
-            
-            func result(for action: KeyboardAction) -> Bool {
-                action.isSystemAction
-            }
             
             it("is true for some actions") {
                 expected = [
@@ -146,8 +143,8 @@ class KeyboardActionTests: QuickSpec {
                     .shift(currentState: .capsLocked),
                     .tab
                 ]
-                expected.forEach { expect(result(for: $0)).to(beTrue()) }
-                unexpected.forEach { expect(result(for: $0)).to(beFalse()) }
+                expected.forEach { expect($0.isSystemAction).to(beTrue()) }
+                unexpected.forEach { expect($0.isSystemAction).to(beFalse()) }
             }
         }
     }